requirements.txt
scrapy
scrapy-playwright
scrapy-redis
celery[redis]
fastapi
uvicorn[standard]
psycopg2-binary
httpx
beautifulsoup4
feedparser
curl_cffi
dateparser==1.2.0
playwright-stealth==1.0.6
jdatetime
pypdf
